+++ ogre2/include/ignition/rendering/ogre2/Ogre2RenderTypes.hh
@@ -32,6 +32,7 @@ namespace ignition
     class Ogre2DepthCamera;
     class Ogre2DirectionalLight;
     class Ogre2Geometry;
+    class Ogre2GizmoVisual;
     class Ogre2GpuRays;
     class Ogre2Light;
     class Ogre2Material;
@@ -66,6 +67,7 @@ namespace ignition
     typedef shared_ptr<Ogre2DepthCamera>          Ogre2DepthCameraPtr;
     typedef shared_ptr<Ogre2DirectionalLight>     Ogre2DirectionalLightPtr;
     typedef shared_ptr<Ogre2Geometry>             Ogre2GeometryPtr;
+    typedef shared_ptr<Ogre2GizmoVisual>          Ogre2GizmoVisualPtr;
     typedef shared_ptr<Ogre2GpuRays>              Ogre2GpuRaysPtr;
     typedef shared_ptr<Ogre2Light>                Ogre2LightPtr;
     typedef shared_ptr<Ogre2Material>             Ogre2MaterialPtr;