+++ ogre2/include/ignition/rendering/ogre2/Ogre2Scene.hh
@@ -119,6 +119,9 @@ namespace ignition
                      const std::string &_name) override;
 
       // Documentation inherited
+      protected: virtual GizmoVisualPtr CreateGizmoVisualImpl(unsigned int _id,
+                     const std::string &_name);
+      // Documentation inherited
       protected: virtual GeometryPtr CreateBoxImpl(unsigned int _id,
                      const std::string &_name) override;