+++ Changelog.md
@@ -4,6 +4,9 @@
 
 ### Ignition Rendering 2.0.0 (20XX-XX-XX)
 
+1. Add GizmoVisual class
+    * [Pull request 163](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/163)
+
 1. Add depth check and depth write properties to Material
     * [Pull request 160](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/160)